refactor: SpaceAccessType -> SpaceVisibility로 이름 변경
