+++ frontend/src/pages/create/funnelElements/spaceVisibilityElement/SpaceVisibilityElement.styles.ts
