+++ frontend/src/types/space.type.ts
@@ -1,2 +1,2 @@
 // TODO: PR별 선언 위치 통합 필요
-export type SpaceAccessType = 'PUBLIC' | 'PRIVATE';
+export type SpaceVisibility = 'PUBLIC' | 'PRIVATE';
